% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPDF.R
\name{findPDF}
\alias{findPDF}
\title{Find PDF files by content with pattern matching}
\usage{
findPDF(path = ".", pattern = "hello world", lowercase = FALSE,
  copy = FALSE, folder = "findPDF", overwrite = TRUE)
}
\arguments{
\item{path}{a character vector, path to be scanned. The default corresponds to the working directory, getwd().}

\item{pattern}{a pattern (regular expression) to search for.}

\item{lowercase}{a logical value. If \code{TRUE}, PDF file content is converted to lowercase before pattern matching.}

\item{copy}{a logical value. If \code{TRUE}, all matching PDF files are copied to \code{folder}.}

\item{folder}{a character vector, path or name of new folder to copy matching PDF files scripts to.}

\item{overwrite}{a logical value. If \code{TRUE}, existing destination files are overwritten.}
}
\description{
\code{findPDF} scans all directories and subdirectories of a given path for PDF files (.pdf) containing
    a specific pattern.  Hits can be copied to a new folder.
}
\examples{
\dontrun{
# Find all of your PDF files containing the word TensorFlow
findPDF(path = "my_files", pattern = "TensorFlow")

# Save the hits to a new folder
findPDF(path = "my_files", pattern = "TensorFlow", copy = TRUE, folder = "pdf_tensorflow")
}
\dontshow{
# Find all PDF files in the package folder that contain the name Hanna
findPDF(path = system.file(package = "findR"), pattern = "Hanna")

# Find all PDF files in the package folder that contain the name Hanna
findPDF(path = system.file(package = "findR"), pattern = "Hanna", copy = TRUE, folder = file.path(tempdir(), "r3"))
list.files(file.path(tempdir(), "r3"))
}
}
