% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findtxt.R
\name{findtxt}
\alias{findtxt}
\title{Find text files by content with pattern matching}
\usage{
findtxt(path = ".", pattern = "hello world", lowercase = FALSE,
  copy = FALSE, folder = "findtxt", overwrite = TRUE)
}
\arguments{
\item{path}{a character vector, path to be scanned. The default corresponds to the working directory, getwd().}

\item{pattern}{a pattern (regular expression) to search for.}

\item{lowercase}{a logical value. If \code{TRUE}, text file content is converted to lowercase before pattern matching.}

\item{copy}{a logical value. If \code{TRUE}, all matching text files are copied to \code{folder}.}

\item{folder}{a character vector, path or name of new folder to copy matching text files to.}

\item{overwrite}{a logical value. If \code{TRUE}, existing destination files are overwritten.}
}
\description{
\code{findtxt} scans all directories and subdirectories of a given path for text files (.txt) containing
    a specific pattern.  Hits can be copied to a new folder.
}
\examples{
\dontrun{
# Find all of your text files containing the word greenhouse
findtxt(path = "my_files", pattern = "greenhouse")

# Save the hits to a new folder
findtxt(path = "my_files", pattern = "greenhouse", copy = TRUE, folder = "txt_greenhouse")
}
\dontshow{
# Find all text files in the package folder that contain the name Einstein
findtxt(path = system.file(package = "findR"), pattern = "einstein")

# Copy the hits to a new folder
findtxt(path = system.file(package = "findR"), pattern = "einstein", copy = TRUE, folder = file.path(tempdir(), "r4"))
list.files(file.path(tempdir(), "r4"))
}
}
