% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPDF.R
\name{findPDF}
\alias{findPDF}
\title{Find PDF files by content with pattern matching}
\usage{
findPDF(path = ".", pattern = "hello world", lowercase = TRUE,
  copy = TRUE, folder = "findPDF", overwrite = TRUE)
}
\arguments{
\item{path}{a character vector, path to be scanned. The default corresponds to the working directory, getwd().}

\item{pattern}{a pattern (regular expression) to search for.}

\item{lowercase}{a logical value. If \code{TRUE}, all PDF file content is converted to lower case.}

\item{copy}{a logical value. If \code{TRUE}, all matching PDF files are copied to \code{folder}.}

\item{folder}{a character vector, path or name of new folder to copy matching R scripts to.}

\item{overwrite}{a logical value. If \code{TRUE}, existing destination files are overwritten.}
}
\description{
\code{findPDF} scans all directories and subdirectories of a given path for PDF files with content
that matches a specific pattern and copies the hits to a folder.
}
\examples{
\dontrun{

# Find all PDF files that contain the name hanna:
findPDF(pattern = "hanna")

}
}
