% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRmd.R
\name{findRmd}
\alias{findRmd}
\title{Find R Markdown files by content with pattern matching}
\usage{
findRmd(path = ".", pattern = "hello world", lowercase = FALSE,
  copy = FALSE, folder = "findRmd", overwrite = TRUE)
}
\arguments{
\item{path}{a character vector, path to be scanned. The default corresponds to the working directory, getwd().}

\item{pattern}{a pattern (regular expression) to search for.}

\item{lowercase}{a logical value. If \code{TRUE}, Rmd file content is converted to lowercase before pattern matching.}

\item{copy}{a logical value. If \code{TRUE}, all matching R scripts are copied to \code{folder}.}

\item{folder}{a character vector, path or name of new folder to copy matching R scripts to.}

\item{overwrite}{a logical value. If \code{TRUE}, existing destination files are overwritten.}
}
\description{
\code{findRmd} scans all directories and subdirectories of a given path for R Markdown files with content
    that matches a specific pattern. Hits can be copied to a new folder.
}
\examples{
\dontrun{
# Find all of your R Markdown files with a ggplot2 scatterplot
findRmd(path = "my_files", pattern = "geom_point")

# Save the hits to a new folder
findRmd(path = "my_files", pattern = "geom_point", copy = TRUE, folder = "Rmd_scatterplot")
}
\dontshow{
# Find all R Markdown files in the package folder that contain a ggplot bar chart
findRmd(path = system.file(package = "findR"), pattern = "geom_bar")

# Copy the hits to a new folder
findRmd(path = system.file(package = "findR"), pattern = "geom_bar", copy = TRUE, folder = file.path(tempdir(), "r2"))
list.files(file.path(tempdir(), "r2"))
}
}
