% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRscript.R
\name{findRscript}
\alias{findRscript}
\title{Find R scripts by content with pattern matching}
\usage{
findRscript(path = ".", pattern = "hello world", lowercase = FALSE,
  comments = TRUE, copy = FALSE, folder = "findRscript",
  overwrite = TRUE)
}
\arguments{
\item{path}{a character vector, path to be scanned. The default corresponds to the working directory, getwd().}

\item{pattern}{a pattern (regular expression) to search for.}

\item{lowercase}{a logical value. If \code{TRUE}, all R script content is converted to lower case.}

\item{comments}{a logical value. If \code{FALSE}, comments (lines starting with #) are excluded from pattern matching.}

\item{copy}{a logical value. If \code{TRUE}, all matching R scripts are copied to \code{folder}.}

\item{folder}{a character vector, path or name of new folder to copy matching R scripts to.}

\item{overwrite}{a logical value. If \code{TRUE}, existing destination files are overwritten.}
}
\description{
\code{findRscript} scans all directories and subdirectories of a given path for R scripts with content
    that matches a specific pattern. Hits can be copied to a new folder.
}
\examples{
\dontrun{
# Find all R scripts in folder 'my_Rscripts' containing a ggplot bar chart:
findRscript(path = "my_Rscripts", pattern = "geom_bar")

# Copy scripts with matching content to a new folder:
findRscript(path = "my_Rscripts", pattern = "geom_bar", copy = T, folder = "my_bar_Rscripts")
}
}
